+++ lib/easy_bills_web/controllers/page_html.ex
@@ -1,7 +1,7 @@
 defmodule EasyBillsWeb.PageHTML do
   use EasyBillsWeb, :html
 
-  alias EasyBillsWeb.IconsComponent
+  alias EasyBillsWeb.CoreComponents
 
   embed_templates "page_html/*"
 end
