+++ lib/easy_bills_web/controllers/page_html/home.html.heex
@@ -13,7 +13,7 @@
     <%= if @current_user do %>
       <div class="space-y-4">
         <div class="flex md:mt-32 lg:mt-0">
-          <IconsComponent.logo_icon />
+          <CoreComponents.logo_icon />
           <h2 class="text-2xl font-bold ml-3 text-purple-500 mt-5">EasyBills</h2>
         </div>
       </div>
@@ -46,7 +46,7 @@
     <% else %>
       <div class="space-y-4">
         <div class="flex md:mt-32 lg:mt-0">
-          <IconsComponent.logo_icon />
+          <CoreComponents.logo_icon />
           <h2 class="text-2xl font-bold ml-3 text-purple-500 mt-5">EasyBills</h2>
         </div>
         <h2 class="mb-8 text-2xl text-cyan-900 font-bold">
