+++ lib/easy_bills_web/live/user_forgot_password_live.ex
@@ -2,7 +2,7 @@ defmodule EasyBillsWeb.UserForgotPasswordLive do
   use EasyBillsWeb, :live_view
 
   alias EasyBills.Accounts
-  alias EasyBillsWeb.IconsComponent
+  alias EasyBillsWeb.CoreComponents
 
   def render(assigns) do
     ~H"""
@@ -16,7 +16,7 @@ defmodule EasyBillsWeb.UserForgotPasswordLive do
       </div>
       <div class="my-auto mx-auto space-y-16 max-w-md">
         <div class="flex mt-16">
-          <IconsComponent.logo_icon />
+          <CoreComponents.logo_icon />
           <h2 class="text-6xl font-bold ml-3 text-purple-600 mt-3">EasyBills</h2>
         </div>
         <.header class="text-center">
