+++ lib/easy_bills_web/live/user_welcome_live.ex
@@ -4,7 +4,7 @@ defmodule EasyBillsWeb.UserWelcomeLive do
   use EasyBillsWeb, :live_view
 
   alias EasyBills.Accounts.User
-  alias EasyBillsWeb.IconsComponent
+  alias EasyBillsWeb.CoreComponents
   # alias EasyBillsWeb.SimpleS3Upload
 
   import Phoenix.HTML.Form
@@ -41,7 +41,7 @@ defmodule EasyBillsWeb.UserWelcomeLive do
       <div class="md:w-[30%] mx-auto mt-16" id="avatar-image-container" phx-hook="UserAvatar">
         <div class="flex mb-14 hidden lg:block">
           <div class="flex">
-            <IconsComponent.logo_icon />
+            <CoreComponents.logo_icon />
             <h2 class="text-6xl font-bold ml-3 text-purple-600 mt-3">EasyBills</h2>
           </div>
         </div>
